Admin: Create a new folder

+++ ContentStorage/Library.cs
@@ -337,11 +337,37 @@ namespace Docomb.ContentStorage
 			if (parent == null) return new(new ActionStatus(ActionStatus.StatusCode.InvalidRequestData, $"No folder was given in which to create a file."), null);
 			string filePath = Path.Combine(parent.FilePath, fileName);
 			if (File.Exists(filePath)) return new(new ActionStatus(ActionStatus.StatusCode.Conflict, $"A file named '{fileName}' already exists."), null);
+			if (Directory.Exists(filePath)) return new(new ActionStatus(ActionStatus.StatusCode.Conflict, $"A folder named '{fileName}' already exists."), null);
 
 			try
 			{
 				File.Create(filePath);
 				if (content?.Length > 0) File.WriteAllText(filePath, content, Encoding.UTF8);
+				List<string> fileParts = new List<string>(parent.UrlParts);
+				fileParts.Add(fileName);
+				ContentItem item = FindItem(fileParts, MatchType.Physical);
+				ClearCache();
+				return new DataWithStatus<ContentItemSummary>(new ActionStatus(ActionStatus.StatusCode.OK), new ContentItemSummary(item));
+			}
+			catch (Exception e)
+			{
+				return new(new ActionStatus(ActionStatus.StatusCode.Error, exception: e), null);
+			}
+		}
+
+		public DataWithStatus<ContentItemSummary> CreateDirectory(ContentDirectory parent, string fileName)
+		{
+			if (string.IsNullOrWhiteSpace(fileName)) return new(new ActionStatus(ActionStatus.StatusCode.InvalidRequestData, "Folder name cannot be empty."), null);
+			if (string.IsNullOrWhiteSpace(GetFileNameWithoutExtension(fileName))) return new(new ActionStatus(ActionStatus.StatusCode.InvalidRequestData, "Folder name cannot be empty."), null);
+			if (Path.GetInvalidFileNameChars().Any(fileName.Contains)) return new(new ActionStatus(ActionStatus.StatusCode.InvalidRequestData, $"New name '{fileName}' contains invalid characters."), null);
+			if (parent == null) return new(new ActionStatus(ActionStatus.StatusCode.InvalidRequestData, $"No folder was given in which to create a sub-folder."), null);
+			string filePath = Path.Combine(parent.FilePath, fileName);
+			if (File.Exists(filePath)) return new(new ActionStatus(ActionStatus.StatusCode.Conflict, $"A file named '{fileName}' already exists."), null);
+			if (Directory.Exists(filePath)) return new(new ActionStatus(ActionStatus.StatusCode.Conflict, $"A folder named '{fileName}' already exists."), null);
+
+			try
+			{
+				Directory.CreateDirectory(filePath);
 				List<string> fileParts = new List<string>(parent.UrlParts);
 				fileParts.Add(fileName);
 				ContentItem item = FindItem(fileParts, MatchType.Physical);
